GeoAPI wrappers for PROJ4J (#115)
* Move to the root `pom.xml` the Maven configuration that was duplicated in modules.
The intend is to avoid more duplication with the introduction of additional modules.
The result should be identical, except the following:
In the `maven-deploy-plugin`, the configuration was:

* <deployAtEnd>true</deployAtEnd> in the `proj4j` module.
* <updateReleaseInfo>true</updateReleaseInfo> in the `epsg` module.

With this commit, the configuration become the following for the two modules:

    <deployAtEnd>true</deployAtEnd>
    <updateReleaseInfo>true</updateReleaseInfo>

Everything else should be effectively identical.

* Move the declaration of JUnit dependency version in the root `pom.xml`,
in order to manage the dependency version in a single place for all modules.

* Add GeoAPI wrappers for GeographicCRS and its dependencies.

* Add wrappers for Projected CRS and Coordinate Operation.
Test the transform of a point from geographic to projected CRS.

* Add wrappers for factories and register them in META-INF/services.
A side-effect of this work is the addition of `proj4j(…)` methods
as the reverse of `geoapi(…)` methods.

* Add a few tests and validations provided by the by GeoAPI conformance module.

* Build with Java 21 with the Java release fixed to 8.
Java version has been verified with `javap`.